[refactor] 클래스 패키지 구조 변경
- 애플리케이션과 테스트 패키지의 클래스들을 목적에 맞게 각각의 패키지로 구분.
