+++ mobile/app/src/main/java/com/lacuc/pets/ui/login/signup/SignUpFragment.kt
@@ -1,4 +1,4 @@
-package com.lacuc.pets.ui
+package com.lacuc.pets.ui.login.signup
 
 import android.os.Bundle
 import android.view.LayoutInflater
