+++ mobile/app/src/main/res/layout/fragment_sign_in.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.SignInFragment">
+    tools:context=".ui.login.signin.SignInFragment">
 
     <ImageView
         android:layout_width="0dp"
