+++ mobile/app/src/main/res/navigation/nav_main.xml
@@ -7,13 +7,13 @@
 
     <fragment
         android:id="@+id/signUpFragment"
-        android:name="com.lacuc.pets.ui.SignUpFragment"
+        android:name="com.lacuc.pets.ui.login.signup.SignUpFragment"
         android:label="회원가입"
         tools:layout="@layout/fragment_sign_up" />
 
     <fragment
         android:id="@+id/signInFragment"
-        android:name="com.lacuc.pets.ui.SignInFragment"
+        android:name="com.lacuc.pets.ui.login.signin.SignInFragment"
         tools:layout="@layout/fragment_sign_in">
         <action
             android:id="@+id/action_signInFragment_to_signUpFragment"
